Map centering
When the map is displayed, we try to centre
on the last user known location.
This simply avoids showing/loading the highest zoom.
(most likely the gps data we are going to show are not so far
from where the user is)

When we get the data we move the camera to include all the data points